requirements.txt
scrapy>=2.11.0
pycryptodome>=3.19.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
